requirements.txt
torch --index-url https://download.pytorch.org/whl/cu121
torchvision --index-url https://download.pytorch.org/whl/cu121
lightning
pyyaml
python-box
pathlib
dvc
dvc-s3
mlflow
dagshub
